Units + White Walkers
Extended the unit selection so a new tier unlocks at mil tech 8
Added a event chain for the White Walkers. Now the summer slowly fades, and once it is over, the White Walkers spawn (with buffs). There is an option for the player to take over the White Walkers when this happens for those that want to play as them.